plat/sgi575: document procedure to setup PXE server
Ubuntu installer allows installation only through CDROM and through
network installer. Since, CDROM is not supported in the SGI-575 model,
the installation can be performed using a network installer. For this
procedure, PXE server has to be setup on the host PC.

This patch adds a document to explain the procedure for setting up the
PXE server. This procedure can be reused for installing other
distributions as well.

Change-Id: I10006048bd58d975ca4fe9af2e4e978fa9ba4a9e
Signed-off-by: Vijayenthiran <vijayenthiran.subramaniam@arm.com>