plat/rdn1edge: document procedure to setup PXE server
Ubuntu installer allows installation only through CDROM and through
network installer. Since, CDROM is not supported in the RD-N1-Edge model,
the installation can be performed using a network installer. For this
procedure, PXE server has to be setup on the host PC.

This patch adds a document to explain the procedure for setting up the
PXE server. This procedure can be reused for installing other
distributions as well.

Change-Id: I2770a323a17f9a9d61059c90215057521f6a2157
Signed-off-by: Vijayenthiran <vijayenthiran.subramaniam@arm.com>
Signed-off-by: Chandni Cherukuri <chandni.cherukuri@arm.com>